requirements.txt
amqp==5.1.1
appdata==2.1.2
asgiref==3.5.2
asyncio==3.4.3
billiard==3.6.4.0
boto3==1.24.83
botocore==1.27.83
celery==5.2.7
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
colored==1.4.3
dataclasses==0.6
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.6
dj-database-url==1.0.0
Django==3.2.14
django-appconf==1.0.5
django-celery==3.1.17
django-crispy-forms==1.13.0
django-encode==1.0.4
django-heroku==0.3.1
django-import-export==2.9.0
django-on-heroku==1.1.2
django-queued-storage==0.8.0
django-ses==3.1.2
django-smtp-ssl==1.0
et-xmlfile==1.1.0
filelock==3.8.0
gunicorn==20.1.0
idna==3.4
jmespath==1.0.1
kombu==5.2.4
MarkupPy==1.14
numpy==1.23.3
odfpy==1.4.1
openpyxl==3.0.10
Pillow==9.2.0
platformdirs==2.5.2
prompt-toolkit==3.0.31
psycopg2==2.9.5
psycopg2-binary==2.9.5
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
razorpay==1.3.0
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
tablib==3.2.1
tzdata==2022.4
urllib3==1.26.12
vine==5.0.0
virtualenv==20.16.5
wcwidth==0.2.5
whitenoise==6.2.0
xlrd==2.0.1
xlwt==1.3.0
